+++ src/generic-table/interfaces/gt-texts.ts
@@ -1,3 +1,5 @@
 export interface GtTexts {
   loading: string;
+  no_data: string;
+  no_matching_data: string;
 }